+++ flowman-kernel-common/src/main/java/com/dimajix/flowman/grpc/GrpcChannel.java
@@ -1,5 +1,0 @@
-package com.dimajix.flowman.grpc;
-
-class GrpcChannel {
-
-}