+++ flowman-kernel-server/src/main/scala/com/dimajix/flowman/kernel/service/SessionManager.scala
@@ -16,6 +16,8 @@
 
 package com.dimajix.flowman.kernel.service
 
+import java.util.UUID
+
 import scala.collection.mutable
 
 import org.slf4j.LoggerFactory
@@ -62,12 +64,12 @@ class SessionManager(val rootSession:execution.Session) {
      * @param projectPath
      * @return
      */
-    def createSession(store: Store, projectName:String) : SessionService = {
+    def createSession(store: Store, projectName:String, clientId:UUID) : SessionService = {
         val project = store.loadProject(projectName)
-        createSession(store, project)
+        createSession(store, project, clientId)
     }
-    def createSession(store: Store, project:Project) : SessionService = {
-        val svc = new SessionService(this, store, project)
+    def createSession(store: Store, project:Project, clientId:UUID) : SessionService = {
+        val svc = new SessionService(this, store, project, clientId)
 
         sessions.synchronized {
             sessions.append(svc)
@@ -75,15 +77,22 @@ class SessionManager(val rootSession:execution.Session) {
 
         svc
     }
-    def createSession(store: Store, projectLocation: File): SessionService = {
+    def createSession(store: Store, projectLocation: File, clientId:UUID): SessionService = {
         if (!projectLocation.isAbsolute())
             throw new IllegalArgumentException(s"Project location is not absolute: $projectLocation")
 
         val project = Project.read.file(projectLocation)
-        createSession(store, project)
+        createSession(store, project, clientId)
     }
-    def createSession(projectLocation:File) : SessionService = {
-        createSession(rootSession.store, projectLocation)
+    def createSession(projectLocation:File, clientId:UUID) : SessionService = {
+        createSession(rootSession.store, projectLocation, clientId)
+    }
+
+    def removeClientSessions(clientId:UUID) : Unit = {
+        val toBeRemoved = sessions.synchronized {
+            sessions.filter(_.clientId == clientId)
+        }
+        toBeRemoved.foreach(_.close())
     }